abstracted plugin class
