+++ dsi/plugins/metadata.py
@@ -1,33 +1,6 @@
 from collections import OrderedDict
-from abc import ABCMeta, abstractmethod
 
-
-class Plugin(metaclass=ABCMeta):
-    """Plugin abstract class for DSI core product.
-
-    A Plugin connects a data writer or reader to a compatible middleware
-    data structure.
-    """
-
-    @abstractmethod
-    def __init__(self, path):
-        """Initialize Plugin setup.
-
-        Read a Plugin file. Return a Plugin object.
-        """
-
-    @property
-    @abstractmethod
-    def git_commit_sha(self):
-        pass
-
-    @abstractmethod
-    def add_to_output(self, path):
-        """Initialize Plugin setup.
-
-        Read a Plugin file. Return a Plugin object.
-        """
-
+from dsi.plugins.plugin import Plugin
 
 class StructuredMetadata(Plugin):
     """ plugin superclass that provides handy methods for structured data """
